增加meta_packet
**现象及复现步骤**

无

**原因和解决方案**

无

**影响范围**

无

**验证方案**

无

**涉及分支**

* master

**检查项**

- [ ] 需要同步修改(trident-dump、trident-copy或trident-testsrv)
- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @zhicong

/cc @xiangyang @yuanchao @jinjie

+++ src/pcap/worker.rs
@@ -158,7 +158,7 @@ impl Worker {
     }
 
     fn get_writer_key(dispatcher_id: u32, acl_gid: u16, tap_type: TapType) -> u64 {
-        (dispatcher_id as u64) << 32 | (acl_gid as u64) << 16 | tap_type as u64
+        (dispatcher_id as u64) << 32 | (acl_gid as u64) << 16 | u16::from(tap_type) as u64
     }
 
     fn finish_writer(writer: Writer, worker_counter_guard: &mut MutexGuard<WorkerCounter>) {
